requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.28
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.3
pydantic>=2.6.4
pydantic-settings>=2.2.1
openai>=1.14.2
boto3>=1.34.69
python-dotenv>=1.0.1
Pillow>=10.2.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.1
requests>=2.31.0
google-genai>=0.1.0
aiosqlite>=0.20.0
